API
Created the API for the resources tab. I added a new folder called screens for the API. This had two typescript folders (Budgeting Advice and Financial Advice) these are connected to the resources.tsx or "Resources" tab. The resources tab now has two buttons that connect to the new screens or layouts. One called Financial Advice, one called Budgeting advice. Each screen is on a stack navigator and are all connected. Financial advice has entries from the strapi under the financial advice category, and budgeting advice has the same but with the budgeting advice category. Each entry is a clickable div with a unique title, description and picture. Clicking it will take the user to the article/information piece.